refactor: Rename /src/pages/Panel/subpanel/ to /components/
BREAKING CHANGE: Renamed some files and components. See details of previous commit and the summary of this commit.
