+++ src/pages/Panel/components/Dictionary.tsx
